+++ client/utilities/ordersUtility.spec.js
@@ -1,7 +1,100 @@
 import { expect } from 'chai';
-import { getOrderStatusDescription, mapOrderSummary, monthlyAverage } from './ordersUtility';
+import {
+  getOrderStatusDescription,
+  mapOrderSummary,
+  monthlyAverage,
+  yearToDateTotal,
+  orderAverage,
+  flattenOrders
+} from './ordersUtility';
 
 describe('Order Status Utility', () => {
+
+  const orders = [
+    {
+      orderID: 954,
+      total: 10.99,
+      createdDate: '2017-04-17T12:11:52.207',
+      orderStatusID: 40,
+      orderStatusDescription: 'Fulfilled',
+      transactionTypeID: 1,
+      totalDiscount: 0.0,
+      totalSub: 9.99,
+      totalTax: 1.0,
+      totalTip: 0.0
+    }, {
+      orderID: 942,
+      total: 20.8725,
+      createdDate: '2017-04-15T22:50:54',
+      orderStatusID: 40,
+      orderStatusDescription: 'Fulfilled',
+      transactionTypeID: 1,
+      totalDiscount: 0.0,
+      totalSub: 18.975,
+      totalTax: 1.8975,
+      totalTip: 0.0
+    }, {
+      orderID: 941,
+      total: 6.6,
+      createdDate: '2017-04-15T22:28:41',
+      orderStatusID: 40,
+      orderStatusDescription: 'Fulfilled',
+      transactionTypeID: 1,
+      totalDiscount: 0.0,
+      totalSub: 6.0,
+      totalTax: 0.6,
+      totalTip: 0.0
+    }, {
+      orderID: 940,
+      total: 0.0,
+      createdDate: '2017-04-15T22:01:02',
+      orderStatusID: 110,
+      orderStatusDescription: 'Refunded',
+      transactionTypeID: 1,
+      totalDiscount: 0.0,
+      totalSub: 0.0,
+      totalTax: 0.0,
+      totalTip: 0.0
+    }, {
+      orderID: 939,
+      total: 0.0,
+      createdDate: '2017-04-15T21:15:07',
+      orderStatusID: 110,
+      orderStatusDescription: 'Refunded',
+      transactionTypeID: 1,
+      totalDiscount: 0.0,
+      totalSub: 0.0,
+      totalTax: 0.0,
+      totalTip: 0.0
+    }, {
+      orderID: 928,
+      total: 0.0,
+      createdDate: '2017-04-15T15:07:37',
+      orderStatusID: 110,
+      orderStatusDescription: 'Refunded',
+      transactionTypeID: 1,
+      totalDiscount: 0.0,
+      totalSub: 0.0,
+      totalTax: 0.0,
+      totalTip: 0.0
+    }, {
+      orderID: 907,
+      total: 9.878,
+      createdDate: '2017-04-14T15:54:42',
+      orderStatusID: 40,
+      orderStatusDescription: 'Fulfilled',
+      transactionTypeID: 1,
+      totalDiscount: 0.0,
+      totalSub: 8.98,
+      totalTax: 0.898,
+      totalTip: 0.0
+    }];
+
+  const results = [
+    { data: orders },
+    { data: orders },
+    { data: orders }
+  ];
 
   const orderData = [{
     createdDate: '2017-02-18T11:30:05.133',
@@ -200,6 +293,24 @@ describe('Order Status Utility', () => {
 
   });
 
+  it('should calculate the average order amount accordingly', () => {
+
+    const expected = 12.09;
+    const actual = orderAverage(orders);
+
+    expect(actual).equals(expected);
+
+  });
+
+  it('should calculate the average order amount as 0 if the list of orders is empty', () => {
+
+    const expected = 0;
+    const actual = orderAverage([]);
+
+    expect(actual).equals(expected);
+
+  });
+
   it('should calculate the monthly average based on the monthly summary passed in', () => {
 
     const expected = 60.38;
@@ -217,8 +328,32 @@ describe('Order Status Utility', () => {
     }];
 
     const actual = monthlyAverage(monthlySummary, property);
+
+    expect(actual).equals(expected);
+
+  });
+
+  it('should flatten orders from the results array', () => {
+
+    const expected = (orders.length) * results.length;
+    const actual = flattenOrders(results)
+      .length;
       
     expect(actual).equals(expected);
+
+  });
+
+  it('should flatten orders from the results array even if the results are all empty objects', () => {
+    const updatedResults = [
+      { data: [] },
+      { data: [] },
+      { data: [] }
+    ];
+
+    const expected = [];
+    const actual = flattenOrders(updatedResults);
+      
+    expect(actual).deep.equals(expected);
 
   });
 
@@ -247,6 +382,66 @@ describe('Order Status Utility', () => {
     const monthlySummary = [];
 
     const actual = monthlyAverage(monthlySummary, property);
+
+    expect(actual).equals(expected);
+
+  });
+
+  it('should calculate the total year to date amount based on the months that are within the current year', () => {
+
+    const expected = 130.76;
+
+    const monthlySummary = [
+      {
+        monthDisplayName: 'Dec',
+        year: 2016,
+        total: 40.38,
+        orderCount: 4
+      },
+      {
+        monthDisplayName: 'Jan',
+        year: (new Date()).getFullYear(),
+        total: 40.38,
+        orderCount: 4
+      },
+      {
+        monthDisplayName: 'Feb',
+        year: (new Date()).getFullYear(),
+        total: 90.38,
+        orderCount: 4
+      }];
+
+    const actual = yearToDateTotal(monthlySummary);
+
+    expect(actual).equals(expected);
+
+  });
+
+  it('should calculate the total year to date amount of 0 based on the months that are within the current year', () => {
+
+    const expected = 0;
+
+    const monthlySummary = [
+      {
+        monthDisplayName: 'Dec',
+        year: 2016,
+        total: 40.38,
+        orderCount: 4
+      },
+      {
+        monthDisplayName: 'Jan',
+        year: 2013,
+        total: 40.38,
+        orderCount: 4
+      },
+      {
+        monthDisplayName: 'Feb',
+        year: 2015,
+        total: 90.38,
+        orderCount: 4
+      }];
+
+    const actual = yearToDateTotal(monthlySummary);
 
     expect(actual).equals(expected);